refactor!: Implement the tmux capturing part in lua instead of bash
This is a breaking change because it removes the 'script_path' setting
and introduces the 'regex' setting instead. If you had a custom script
before, it means that you have to remove it and instead put your regex
in the config options.

This makes it easier to update the regex if needed (which hopefully
should be very rare).